# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorDto": {
        "properties": {
          "code": {
            "example": "VALIDATION_FAILED",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "type": "object"
          },
          "message": {
            "example": "请求参数验证失败",
            "type": "string"
          },
          "requestId": {
            "example": "7da2aa93-ef82-45c8-9df0-84232e6a5b13",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "details"
        ],
        "type": "object"
      },
...
      "CreateProductRequestDto": {
        "properties": {
          "description": {
            "maxLength": 5000,
            "minLength": 1,
            "type": "string"
          },
          "imageKey": {
            "maxLength": 200,
            "pattern": "^products/[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}\\.(jpg|png|webp)$",
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "pointsCost": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "number"
          },
          "stock": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "name",
          "description",
          "imageKey",
          "stock",
          "pointsCost"
        ],
        "type": "object"
      },
...
      "PageMetaDto": {
        "properties": {
          "page": {
            "example": 1,
            "minimum": 1,
            "type": "number"
          },
          "pageSize": {
            "example": 20,
            "maximum": 100,
            "minimum": 1,
            "type": "number"
          },
          "total": {
            "example": 42,
            "minimum": 0,
            "type": "number"
          },
          "totalPages": {
            "example": 3,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "type": "object"
      },
...
      "ProductDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "imageKey": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "pointsCost": {
            "minimum": 0,
            "type": "number"
          },
          "stock": {
            "minimum": 0,
            "type": "number"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "imageKey",
          "stock",
          "pointsCost",
          "isActive",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "ProductImageUploadResponseDto": {
        "properties": {
          "key": {
            "example": "products/550e8400-e29b-41d4-a716-446655440000.png",
            "type": "string"
          },
          "url": {
            "example": "/uploads/products/550e8400-e29b-41d4-a716-446655440000.png",
            "type": "string"
          }
        },
        "required": [
          "key",
          "url"
        ],
        "type": "object"
      },
      "ProductListResponseDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "UpdateProductRequestDto": {
        "properties": {
          "description": {
            "maxLength": 5000,
            "minLength": 1,
            "type": "string"
          },
          "imageKey": {
            "maxLength": 200,
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "pointsCost": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "number"
          },
          "stock": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "number"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Android 与其他无 Cookie 客户端使用的访问令牌",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "description": "Web 客户端访问令牌 Cookie",
        "in": "cookie",
        "name": "pq_access",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "英语答题、积分、商品兑换与订单管理开放接口",
    "title": "Point Quest API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/products": {
      "get": {
        "operationId": "adminListProducts",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "isActive",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000000,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponseDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "分页查询全部商品",
        "tags": [
          "管理端-商品"
        ]
      },
      "post": {
        "operationId": "adminCreateProduct",
        "parameters": [
          {
            "description": "使用 Cookie 身份认证执行写操作时必填；Bearer 模式勿填",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "创建商品",
        "tags": [
          "管理端-商品"
        ]
      }
    },
    "/api/v1/admin/products/{productId}": {
      "patch": {
        "operationId": "adminUpdateProduct",
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "maxLength": 191,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "使用 Cookie 身份认证执行写操作时必填；Bearer 模式勿填",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "更新商品、库存或上下架状态",
        "tags": [
          "管理端-商品"
        ]
      }
    },
...
    "/api/v1/admin/uploads/product-images": {
      "post": {
        "operationId": "adminUploadProductImage",
        "parameters": [
          {
            "description": "使用 Cookie 身份认证执行写操作时必填；Bearer 模式勿填",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "description": "JPEG、PNG 或 WebP，最大 5 MiB、2500 万像素、单帧",
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageUploadResponseDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "上传并规范化商品图片",
        "tags": [
          "管理端-上传"
        ]
      }
    },
...
    "/api/v1/products/{productId}": {
      "get": {
        "operationId": "productsGet",
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "maxLength": 191,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "获取可兑换商品详情",
        "tags": [
          "学生端-商品"
        ]
      }
    }
  },
  "servers": [],
...
}